move game management to context provider
